攏khioh去匯入。
